backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
websockets==12.0

# Pydantic
pydantic==2.9.0
pydantic-settings==2.4.0

# Jupyter kernel communication
jupyter-client==8.6.2
ipykernel==6.29.5
pyzmq==26.1.0

# AI Providers
anthropic==0.34.0
openai==1.45.0
google-generativeai==0.7.2

# Async support
aiofiles==24.1.0
aiohttp==3.10.5

# Database (optional)
sqlalchemy==2.0.32
aiosqlite==0.20.0

# Redis (optional)
redis==5.0.8

# Utils
python-dotenv==1.0.1
